requirements.txt
torch
torchaudio
so-vits-svc-fork
numpy
noisereduce
pydub
scipy
ffmpeg-python
fastapi
uvicorn
openai-whisper
edge-tts
deep-translator